refactor: rename HttpInterfaceConfig to add an interface to it
